# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/postgres/users": {
            "post": {
                "description": "Добавляет нового пользователя в PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать пользователя в БД",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID нового пользователя",
                        "schema": {
                            "$ref": "#/definitions/service.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/redis/incr": {
            "post": {
                "description": "Увеличивает значение по ключу на переданное значение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Инкрементировать значение в Redis",
                "parameters": [
                    {
                        "description": "Данные запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.RequestPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ с новым значением",
                        "schema": {
                            "$ref": "#/definitions/service.ResponsePayload"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sign/hmacsha512": {
            "post": {
                "description": "Возвращает HMAC-SHA512 подпись переданного текста с указанным ключом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Генерация HMAC-SHA512 подписи",
                "parameters": [
                    {
                        "description": "Данные запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.HMACRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ с подписью",
                        "schema": {
                            "$ref": "#/definitions/service.HMACResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.HMACRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "service.HMACResponse": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string"
                }
            }
        },
        "service.RequestPayload": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "service.ResponsePayload": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "service.UserRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}